backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
asyncmy
greenlet
python-dotenv
pyjwt
openpyxl
minio
python-multipart